add mongodb dst

+++ scripts/menu.sh
@@ -60,9 +60,11 @@ copy_yaml() {
     local SRCDB_DIR=$1
     local DSTDB_DIR=$2
     local DSTDB_GRP=$3
+    local PID
 
     # copy the base src and dir config
     for f in $SCRIPTS_DIR/$SRCDB_DIR/src*.yaml $SCRIPTS_DIR/$DSTDB_DIR/dst*.yaml; do
+        echo copy $f
         cat $f | PID=$$ envsubst > $CFG_DIR/$(basename $f) 
     done
 
@@ -75,10 +77,10 @@ copy_yaml() {
     for f in $CFG_DIR/*.yaml; do
         filename=$( basename $f )   
 
-        if [ -f "$SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$DSTDB_DIR/$filename" ]; then
+        if [ ! -z "$DSTDB_GRP" -a ! -z "$DSTDB_DIR" -a -f "$SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$DSTDB_DIR/$filename" ]; then
             echo override cat $SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$DSTDB_DIR/$filename \| PID=$$ envsubst \> $CFG_DIR/$filename
             cat $SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$DSTDB_DIR/$filename | PID=$$ envsubst > $CFG_DIR/$filename
-        elif [ -f "$SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$filename" ]; then
+        elif [ ! -z "$DSTDB_GRP" -a -f "$SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$filename" ]; then
             echo override cat $SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$filename \| PID=$$ envsubst \> $CFG_DIR/$filename
             cat $SCRIPTS_DIR/$SRCDB_DIR/$DSTDB_GRP/$filename | PID=$$ envsubst > $CFG_DIR/$filename
         fi
@@ -253,7 +255,9 @@ ask_repl_mode() {
     export REPL_TYPE
 }
 init_src() {
-    rc=0
+    local DB_INIT
+    local DB_GRP
+    local rc=0
     # find src.init.sh
     if [ -f "${SCRIPTS_DIR}/${SRCDB_DIR}/src.init.sh" ]; then
         DB_INIT=${SCRIPTS_DIR}/${SRCDB_DIR}/src.init.sh
@@ -267,23 +271,27 @@ init_src() {
     if [ -f "${DB_INIT}" ]; then
             # NOTE: do not remove () below as that will exit this script
             ( exec ${DB_INIT} ) 
-            if [ ! -z "$( cat $CFG_DIR/src.init.sh.log | grep -i failed )" ]; then rc=1; fi  
+            if [  ! -z "$DB_GRP" -a ! -z "$( cat $CFG_DIR/src.init.sh.log | grep -i failed )" ]; then rc=1; fi  
     else
         echo "${SCRIPTS_DIR}/${SRCDB_DIR}/src.init.sh: not found. skipping"    
     fi
     return $rc
 }
 init_dst() {
-    rc=0
+    local DB_INIT
+    local DB_GRP
+    local rc=0    
     # find dst.init.sh
     if [ -f "${SCRIPTS_DIR}/${DSTDB_DIR}/dst.init.sh" ]; then
         DB_INIT=${SCRIPTS_DIR}/${DSTDB_DIR}/dst.init.sh
     elif [ -f "${SCRIPTS_DIR}/utils/map.csv" ]; then 
         DB_GRP=$(grep "^${DSTDB_DIR}," ${SCRIPTS_DIR}/utils/map.csv | cut -d',' -f2)
-        if [ -f "${SCRIPTS_DIR}/utils/${DB_GRP}/dst.init.sh" ]; then
+        if [ ! -z "$DB_GRP" -a -f "${SCRIPTS_DIR}/utils/${DB_GRP}/dst.init.sh" ]; then
             DB_INIT=${SCRIPTS_DIR}/utils/${DB_GRP}/dst.init.sh
+            echo DB_INIT=${SCRIPTS_DIR}/utils/${DB_GRP}/dst.init.sh
         fi
     fi
+
     # run dst.init.sh
     if [ -f "${DB_INIT}" ]; then
             # NOTE: do not remove () below as that will exit this script
